requirements.txt
appnope==0.1.2
asgiref==3.3.1
backcall==0.2.0
boto3==1.17.1
botocore==1.20.1
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
Django==3.1.6
django-storages==1.11.1
djangorestframework==3.12.2
drf-nested-routers==0.92.5
idna==2.10
jedi==0.18.0
jmespath==0.10.0
oauthlib==3.1.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.14
ptyprocess==0.7.0
Pygments==2.7.4
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
tqdm==4.56.0
urllib3==1.26.3
wcwidth==0.2.5
